requirements.txt
aiofile==3.8.8
pyaudio==0.2.13
amazon-transcribe==0.6.2
sounddevice==0.4.6
numpy==1.26.0
# pyannote.audio==2.1.1
cohere==4.27
python-dotenv==1.0.0
openai==0.28.0
google-cloud-speech==2.21.0
pyautogui==0.9.54
scipy==1.11.2
soundfile==0.12.1
supabase==1.0.4
streamlit==1.27.0
